Update 编程填空：学生信息处理程序.cpp
经过修改，将 平均成绩 和 各次成绩 改为了浮点类型。

+++ 编程填空：学生信息处理程序.cpp
@@ -12,14 +12,14 @@ private:
 	string name;
 	int age;
 	int num;
-	int average_grade;
-	int grades[4];
+	float average_grade;
+	float grades[4];
 public:
 	void input() {
 		getline(cin, name, ',');
 		char ch;
 		cin >> age >> ch >> num;
-		int a = 0;
+		float a = 0;
 		for (int i = 0; i < 4; i++) {
 			cin >> ch >> a;
 			grades[i] = a;
@@ -27,7 +27,7 @@ public:
 	}
 
 	void calculate() {
-		int sum = 0;
+		float sum = 0;
 		for (int i = 0; i < 4; i++) {
 			sum += grades[i];
 		}
